Fixed batch cov string

+++ neuroCombat/neuroCombat.py
@@ -104,7 +104,7 @@ def neuroCombat(dat,
     # create dictionary that stores batch info
     (batch_levels, sample_per_batch) = np.unique(covars[:,batch_col],return_counts=True)
     info_dict = {
-        'batch_levels': batch_levels.astype('int'),
+        'batch_levels': batch_levels,
         'ref_level': ref_level,
         'n_batch': len(batch_levels),
         'n_sample': int(covars.shape[0]),
